+++ docs/changelog.rst
@@ -12,8 +12,9 @@ Version 1.5.2
 
 **Unreleased**
 
+- Forbid fragments in ``redirect_uris``. :issue:`714`
 - Fix invalid characters in ``error_description``. :issue:`720`
-- Add ``claims_cls``` parameter for client's ``parse_id_token`` method.
+- Add ``claims_cls``` parameter for client's ``parse_id_token`` method. :issue:`725`
 
 
 Version 1.5.1